--- eShopModernizedWebFormsSolution/src/eShopModernizedWebForms/About.aspx.cs
@@ -1,16 +1,0 @@
-﻿using log4net;
-using System;
-using System.Web.UI;
-
-namespace eShopModernizedWebForms
-{
-    public partial class About : Page
-    {
-        private static readonly ILog _log = LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
-
-        protected void Page_Load(object sender, EventArgs e)
-        {
-            _log.Info($"Now loading... /About.aspx");
-        }
-    }
-}
