+++ eShopModernizedWebFormsSolution/src/eShopModernizedWebForms/App_Start/MyTelemetryInitializer.cs
